Rename toast directory
